src/advertising/requirements.txt
Flask
flask_sqlalchemy
flask_migrate
wtforms
flask_wtf
requests
